KAMP univariate Expectation (Matrix Implementation)

Computes the KAMP (K-function Adjusted for Marked Permutations) expectation for a given spatial point pattern. This function calculates Ripley's K using both the traditional Ripley's K method (based on Kcross) and the KAMP-adjusted CSR baseline (based on Kest).

The KAMP-adjusted CSR represents a more realistic baseline for K (compared to traditional CSR) that accounts for spatial clustering or inhomogeneity in a point pattern compared to the traditional CSR assumption, while avoiding the computational burden of permuting the point pattern.

Note: this is a slower, matrix-based implementation of the KAMP expectation

See get_kamp_expectation for the faster implementation that utilizes functions in spatstat

Usage

kamp_expectation_mat(
  ppp_obj,
  rvec = c(0, 0.05, 0.075, 0.1, 0.15, 0.2),
  correction = "trans",
  markvar = "immune",
  thin_pct = 0
)

Arguments

ppp_obj

A point pattern object of class "ppp" from the spatstat package.

rvec

A vector of radii at which to calculate the KAMP expectation.

correction

Type of edge correction. Defaults to translational.

markvar

The variable used to mark the points in the point pattern object. Defaults to "immune".

thin_pct

Percentage that determines how much to thin the amount of points in the point pattern object. Defaults to 0.

Value

A dataframe with the following columns:

r

The radius at which K was calculated.

k

The observed K value calculated using matrices

theo_csr

The theoretical K under CSR

kamp_csr

The adjusted CSR representing the permuted expectation.

kamp_fundiff

The difference between observed K and KAMP CSR

Details

Computes KAMP Expectation using matrices

Examples

if (requireNamespace("spatstat.geom", quietly = TRUE)) {
# simulate a simple spatial point pattern with two types
win <- spatstat.geom::owin(c(0, 1), c(0, 1))
pp <- spatstat.random::rpoispp(lambda = 100, win = win)
marks <- sample(c("immune", "background"), pp$n, replace = TRUE)
marked_pp <- spatstat.geom::ppp(pp$x, pp$y, window = win, marks = factor(marks))
# compute KAMP expectation using matrix method
kamp_result_mat <- kamp_expectation_mat(marked_pp, markvar = "immune")
print(kamp_result_mat)
}
#> # A tibble: 6 × 5
#>       r       k theo_csr kamp_csr kamp_fundiff
#>   <dbl>   <dbl>    <dbl>    <dbl>        <dbl>
#> 1 0     0        0        0            0      
#> 2 0.05  0.00600  0.00785  0.00731     -0.00131
#> 3 0.075 0.0131   0.0177   0.0153      -0.00219
#> 4 0.1   0.0269   0.0314   0.0297      -0.00286
#> 5 0.15  0.0597   0.0707   0.0703      -0.0106 
#> 6 0.2   0.117    0.126    0.122       -0.00485